requirements.txt
torch
pandas
numpy
scikit-learn
fpdf2
matplotlib
joblib